+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Precise UI Changelog
 
-## 2.0.1
+## 2.1.0
 
+- Flyout component reimplemented using Popper.js (556349)
+- HOC for using functional components with react-onclickoutside implemented
 - Fix onChangeRow() method in DateField component
 - Use modern Popper modifiers format in DateField component